feat: update search command
